#1063 - Edited stomp-realted integration tests.
 #1063 - Updated github action to reflect rename of stomp to stomp_ in integration tests.

 #1063 - Removed unused imports and reorganized the remaining via isort.

 #1063 - Edited except clause to satisfy flake8.
